refactor: package RPC endpoints together
