+++ arithmetization/src/main/java/net/consensys/linea/services/tracegeneration/rpc/FileTrace.java
@@ -13,7 +13,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.tracegeneration.rpc;
+package net.consensys.linea.services.tracegeneration.rpc;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
